+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
 
 ### Changed
 - Create/Update element actions are specific to elements
+- CheckAccess trait now throws an UnauthorizedHttpException
 
 ## 1.0.0-rc.7 - 2018-01-17
 ### Changed